Don't create/update group objects in Accounts if the requestor is temp

+++ app/models/openstax/accounts/group.rb
@@ -115,21 +115,21 @@ module OpenStax::Accounts
     end
 
     def create_openstax_accounts_group
-      return false unless requestor
+      return false if requestor.nil? || requestor.is_anonymous?
 
-      OpenStax::Accounts::Api.create_group(requestor, self)
+      OpenStax::Accounts::Api.create_group(requestor, self) if requestor.has_authenticated?
     end
 
     def update_openstax_accounts_group
-      return false unless requestor
+      return false if requestor.nil? || requestor.is_anonymous?
 
-      OpenStax::Accounts::Api.update_group(requestor, self)
+      OpenStax::Accounts::Api.update_group(requestor, self) if requestor.has_authenticated?
     end
 
     def destroy_openstax_accounts_group
-      return false unless requestor
+      return false if requestor.nil? || requestor.is_anonymous?
 
-      OpenStax::Accounts::Api.destroy_group(requestor, self)
+      OpenStax::Accounts::Api.destroy_group(requestor, self) if requestor.has_authenticated?
     end
 
   end
