+++ app/models/openstax/accounts/group_member.rb
@@ -22,15 +22,15 @@ module OpenStax::Accounts
     end
 
     def create_openstax_accounts_group_member
-      return false unless requestor
+      return false if requestor.nil? || requestor.is_anonymous?
 
-      OpenStax::Accounts::Api.create_group_member(requestor, self)
+      OpenStax::Accounts::Api.create_group_member(requestor, self) if requestor.has_authenticated?
     end
 
     def destroy_openstax_accounts_group_member
-      return false unless requestor
+      return false if requestor.nil? || requestor.is_anonymous?
 
-      OpenStax::Accounts::Api.destroy_group_member(requestor, self)
+      OpenStax::Accounts::Api.destroy_group_member(requestor, self) if requestor.has_authenticated?
     end
 
   end
